proxy: Prepare for multiple kinds of proxies.
Relocate L7 rule translation from policy package to proxy package and
move all oxy code to a new file oxyproxy.go in preparation to
supporting multiple kinds of proxies.

As part of this change the type policy.AuxRule is eliminated, and the
api.L7Rules is used instead, so that the proxy specific code can
translate the l7 rules.

With this change each redirection is implemented through an Redirect
interface, each of which can be backed by different proxy
implementations.

Policy code is changed to handle conflicts in L4 policy merges due to
conflicting parser types, redirect port values and conflicting l7
rules.
