+++ pkg/policy/l4.go
@@ -24,11 +24,6 @@ import (
 	"github.com/cilium/cilium/pkg/policy/api"
 )
 
-type AuxRule struct {
-	Expr     string `json:"expr"`
-	L7Parser string `json:"l7Parser"`
-}
-
 // L7ParserType is the type used to indicate what L7 parser to use and
 // defines all supported types of L7 parsers
 type L7ParserType string
@@ -46,105 +41,13 @@ type L4Filter struct {
 	// Protocol is the L4 protocol to allow or NONE
 	Protocol string
 	// L7Parser specifies the L7 protocol parser (optional)
-	L7Parser string
+	L7Parser L7ParserType
 	// L7RedirectPort is the L7 proxy port to redirect to (optional)
 	L7RedirectPort int
 	// L7Rules is a list of L7 rules which are passed to the L7 proxy (optional)
-	L7Rules []AuxRule
+	L7Rules api.L7Rules
 	// Ingress is true if filter applies at ingress
 	Ingress bool
-}
-
-// FillPortRuleHTTP fills in the http port rule fields
-// as a regex into a single string
-func FillPortRuleHTTP(httpRule []api.PortRuleHTTP) ([]AuxRule, string) {
-	l7rules := []AuxRule{}
-	l7Parser := ""
-	for _, h := range httpRule {
-		r := AuxRule{L7Parser: string(ParserTypeHTTP)}
-
-		if h.Path != "" {
-			r.Expr = "PathRegexp(\"" + h.Path + "\")"
-		}
-
-		if h.Method != "" {
-			if r.Expr != "" {
-				r.Expr += " && "
-			}
-			r.Expr += "MethodRegexp(\"" + h.Method + "\")"
-		}
-
-		if h.Host != "" {
-			if r.Expr != "" {
-				r.Expr += " && "
-			}
-			r.Expr += "HostRegexp(\"" + h.Host + "\")"
-		}
-
-		for _, hdr := range h.Headers {
-			s := strings.SplitN(hdr, " ", 2)
-			if r.Expr != "" {
-				r.Expr += " && "
-			}
-			r.Expr += "Header(\""
-			if len(s) == 2 {
-				// Remove ':' in "X-Key: true"
-				key := strings.TrimRight(s[0], ":")
-				r.Expr += key + "\",\"" + s[1]
-			} else {
-				r.Expr += s[0]
-			}
-			r.Expr += "\")"
-		}
-
-		if r.Expr != "" {
-			l7rules = append(l7rules, r)
-		}
-	}
-
-	if len(l7rules) > 0 {
-		l7Parser = string(ParserTypeHTTP)
-	}
-	return l7rules, l7Parser
-}
-
-// FillPortRuleKafka fills in the Kafka port rule fields
-// as a regex into a single string
-// TODO The string formed is likely not the final format, and will
-// evolve as more kafka keys are supported.
-func FillPortRuleKafka(kafkaRule []api.PortRuleKafka) ([]AuxRule, string) {
-	l7rules := []AuxRule{}
-	l7Parser := ""
-	for _, h := range kafkaRule {
-		r := AuxRule{L7Parser: string(ParserTypeKafka)}
-
-		if h.APIVersion != "" {
-			r.Expr = "ApiVersionRegexp(\"" + h.APIVersion + "\")"
-		}
-
-		if h.Topic != "" {
-			if r.Expr != "" {
-				r.Expr += " && "
-			}
-			r.Expr += "TopicRegexp(\"" + h.Topic + "\")"
-		}
-
-		if h.APIKey != "" {
-			if r.Expr != "" {
-				r.Expr += " && "
-			}
-			r.Expr += "ApiKeyRegexp(\"" + h.APIKey + "\")"
-		}
-
-		if r.Expr != "" {
-			l7rules = append(l7rules, r)
-		}
-	}
-
-	if len(l7rules) > 0 {
-		l7Parser = string(ParserTypeKafka)
-	}
-	return l7rules, l7Parser
 }
 
 // CreateL4Filter creates an L4Filter based on an api.PortRule and api.PortProtocol
@@ -163,12 +66,12 @@ func CreateL4Filter(rule api.PortRule, port api.PortProtocol, direction string, 
 	}
 
 	if rule.Rules != nil {
-		switch {
-		case rule.Rules.HTTP != nil:
-			l4.L7Rules, l4.L7Parser = FillPortRuleHTTP(rule.Rules.HTTP)
-		case rule.Rules.Kafka != nil:
-			l4.L7Rules, l4.L7Parser = FillPortRuleKafka(rule.Rules.Kafka)
+		if len(rule.Rules.HTTP) > 0 {
+			l4.L7Parser = ParserTypeHTTP
+		} else if len(rule.Rules.Kafka) > 0 {
+			l4.L7Parser = ParserTypeKafka
 		}
+		l4.L7Rules = *rule.Rules
 	}
 
 	return l4
